+++ Source/Testably.Abstractions.Testing/FileSystemMock.CallbackHandler.cs
@@ -8,19 +8,19 @@ public sealed partial class FileSystemMock
 {
     internal sealed class FileSystemMockCallbackHandler : ICallbackHandler
     {
-        private readonly Notification.INotificationFactory<FileSystemChange>
-            _changeOccurringCallbacks = Notification.CreateFactory<FileSystemChange>();
+        private readonly Notification.INotificationFactory<CallbackChange>
+            _changeOccurringCallbacks = Notification.CreateFactory<CallbackChange>();
 
-        private readonly Notification.INotificationFactory<FileSystemChange>
-            _changeOccurredCallbacks = Notification.CreateFactory<FileSystemChange>();
+        private readonly Notification.INotificationFactory<CallbackChange>
+            _changeOccurredCallbacks = Notification.CreateFactory<CallbackChange>();
 
         #region ICallbackHandler Members
 
         /// <inheritdoc
         ///     cref="FileSystemMock.ICallbackHandler.ChangeOccurring(Action{ICallbackHandler.FileSystemChange}, Func{ICallbackHandler.FileSystemChange,bool})" />
         public ICallbackHandler ChangeOccurring(
-            Action<FileSystemChange> callback,
-            Func<FileSystemChange, bool>? predicate = null)
+            Action<CallbackChange> callback,
+            Func<CallbackChange, bool>? predicate = null)
         {
             _changeOccurringCallbacks.RegisterCallback(callback, predicate);
             return this;
@@ -28,23 +28,23 @@ public sealed partial class FileSystemMock
 
         /// <inheritdoc
         ///     cref="FileSystemMock.ICallbackHandler.ChangeOccurred(Action{ICallbackHandler.FileSystemChange}, Func{ICallbackHandler.FileSystemChange,bool})" />
-        public Notification.IAwaitableCallback<FileSystemChange> ChangeOccurred(
-            Action<FileSystemChange>? callback = null,
-            Func<FileSystemChange, bool>? predicate = null)
+        public Notification.IAwaitableCallback<CallbackChange> ChangeOccurred(
+            Action<CallbackChange>? callback = null,
+            Func<CallbackChange, bool>? predicate = null)
             => _changeOccurredCallbacks.RegisterCallback(callback, predicate);
 
         #endregion
 
-        public FileSystemChange InvokeChangeOccurring(string path,
+        public CallbackChange InvokeChangeOccurring(string path,
                                                       CallbackChangeType changeType,
                                                       NotifyFilters notifyFilters)
         {
-            FileSystemChange fileSystemChange = new(path, changeType, notifyFilters);
+            CallbackChange fileSystemChange = new(path, changeType, notifyFilters);
             _changeOccurringCallbacks.InvokeCallbacks(fileSystemChange);
             return fileSystemChange;
         }
 
-        public void InvokeChangeOccurred(FileSystemChange? fileSystemChange)
+        public void InvokeChangeOccurred(CallbackChange? fileSystemChange)
         {
             if (fileSystemChange != null)
             {
